requirements.txt
python-slugify
rdflib
requests
sparqlwrapper
